smith-api-poc/requirements.txt
requests
python-dotenv
autopep8